backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
supabase==2.3.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai>=0.3.2
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
httpx[http2]>=0.24.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
